requirements.txt
pytorch_lightning==2.5.2
kaldialign==0.9.2
onnx==1.18.0
onnxruntime-gpu==1.22.0
